relocate transcripts from solver to log module
