+++ src/inspect_ai/log/__init__.py
@@ -23,6 +23,7 @@ from ._log import (
 )
 from ._message import LoggingLevel, LoggingMessage
 from ._retry import retryable_eval_logs
+from ._transcript import Transcript, transcript
 
 __all__ = [
     "EvalConfig",
@@ -41,6 +42,8 @@ __all__ = [
     "EvalLogInfo",
     "LoggingLevel",
     "LoggingMessage",
+    "Transcript",
+    "transcript",
     "list_eval_logs",
     "read_eval_log",
     "write_eval_log",
