+++ src/inspect_ai/log/_transcript.py
@@ -30,8 +30,8 @@ from inspect_ai.tool._tool_call import ToolCallError
 from inspect_ai.tool._tool_choice import ToolChoice
 from inspect_ai.tool._tool_info import ToolInfo
 
-from .._task_state import state_jsonable
-from .store import store, store_changes, store_jsonable
+from ..solver._subtask.store import store, store_changes, store_jsonable
+from ..solver._task_state import state_jsonable
 
 logger = getLogger(__name__)
