+++ src/inspect_ai/solver/__init__.py
@@ -11,7 +11,6 @@ from ._prompt import (
 from ._solver import Generate, Solver, generate, solver
 from ._subtask.store import Store, store
 from ._subtask.subtask import Subtask, subtask
-from ._subtask.transcript import Transcript, transcript
 from ._task_state import Choice, Choices, TaskState
 from ._use_tools import use_tools
 
@@ -33,8 +32,6 @@ __all__ = [
     "Generate",
     "Store",
     "store",
-    "Transcript",
-    "transcript",
     "Subtask",
     "subtask",
 ]
@@ -42,6 +39,7 @@ __all__ = [
 
 _TOOL_MODULE_VERSION_3_18 = "0.3.18"
 _TOOL_MODULE_VERSION_3_19 = "0.3.19"
+_SUBTASKS_MODULE_VERSION = "0.3.26"
 _REMOVED_IN = "0.4"
 relocated_module_attribute(
     "Tool", "inspect_ai.tool.Tool", _TOOL_MODULE_VERSION_3_18, _REMOVED_IN
@@ -90,4 +88,16 @@ relocated_module_attribute(
 )
 relocated_module_attribute(
     "web_search", "inspect_ai.tool.web_search", _TOOL_MODULE_VERSION_3_19, _REMOVED_IN
+)
+relocated_module_attribute(
+    "Transcript",
+    "inspect_ai.log.Transcript",
+    _SUBTASKS_MODULE_VERSION,
+    _REMOVED_IN,
+)
+relocated_module_attribute(
+    "transcript",
+    "inspect_ai.log.transcript",
+    _SUBTASKS_MODULE_VERSION,
+    _REMOVED_IN,
 )
